First 3 methods
What's up guys, I put my ideas for the first three methods commented out in the code. I didn't want to change what was already there, but I think the current loader will cause every book to be empty, so my idea was to create a separate ArrayList for each book that gets added to loadedData.